Enemy units now inherit from military units and 2 new enemy types have been made
