+++ Singularity/Singularity/Screen/ScreenClasses/GameScreen.cs
@@ -124,7 +124,7 @@ namespace Singularity.Screen.ScreenClasses
                 var possiblegenunit = drawable as GeneralUnit;
                 var possiblerock = drawable as Rock;
                 var possiblepuddle = drawable as Puddle;
-                var conUnit = drawable as ControllableUnit;
+                var conUnit = drawable as FreeMovingUnit;
                 if (conUnit != null)
                 {
                     mSelBox.SelectingBox += conUnit.BoxSelected;
@@ -150,10 +150,10 @@ namespace Singularity.Screen.ScreenClasses
                 var possiblegenunit = updateable as GeneralUnit;
                 var possiblerock = updateable as Rock;
                 var possiblepuddle = updateable as Puddle;
-                var conUnit = updateable as ControllableUnit;
-                if (conUnit != null)
+                var freeMovingUnit = updateable as FreeMovingUnit;
+                if (freeMovingUnit != null && freeMovingUnit.Friendly)
                 {
-                    mSelBox.SelectingBox += conUnit.BoxSelected;
+                    mSelBox.SelectingBox += freeMovingUnit.BoxSelected;
                 }
                 possiblepuddle?.ReloadContent();
                 possiblerock?.ReloadContent();
@@ -176,10 +176,10 @@ namespace Singularity.Screen.ScreenClasses
                 var possiblegenunit = spatial as GeneralUnit;
                 var possiblerock = spatial as Rock;
                 var possiblepuddle = spatial as Puddle;
-                var conUnit = spatial as ControllableUnit;
-                if (conUnit != null)
+                var freeMovingUnit = spatial as FreeMovingUnit;
+                if (freeMovingUnit != null && freeMovingUnit.Friendly)
                 {
-                    mSelBox.SelectingBox += conUnit.BoxSelected;
+                    mSelBox.SelectingBox += freeMovingUnit.BoxSelected;
                 }
                 possiblepuddle?.ReloadContent();
                 possiblerock?.ReloadContent();
@@ -332,8 +332,7 @@ namespace Singularity.Screen.ScreenClasses
             var road = toAdd as Road;
             var platform = toAdd as PlatformBlank;
             var settler = toAdd as Settler;
-            var conUnit = toAdd as ControllableUnit;
-            var enemyUnit = toAdd as EnemyUnit; // currently unnecessary
+            var freeMovingUnit = toAdd as FreeMovingUnit;
 
             if (!typeof(IDraw).IsAssignableFrom(typeof(T)) && !typeof(IUpdate).IsAssignableFrom(typeof(T)) && road == null && platform == null)
             {
@@ -363,15 +362,14 @@ namespace Singularity.Screen.ScreenClasses
             }
 
             // subscribe every military unit to the selection box
-            if (conUnit != null)
+            if (freeMovingUnit != null)
             {
-                mSelBox.SelectingBox += conUnit.BoxSelected;
-                mDirector.GetMilitaryManager.AddUnit(conUnit);
-            }
+                if (freeMovingUnit.Friendly)
+                {
+                    mSelBox.SelectingBox += freeMovingUnit.BoxSelected;
+                }
 
-            if (enemyUnit != null)
-            {
-                mDirector.GetMilitaryManager.AddUnit(enemyUnit);
+                mDirector.GetMilitaryManager.AddUnit(freeMovingUnit);
             }
 
             if (typeof(IRevealing).IsAssignableFrom(typeof(T)))
@@ -427,7 +425,7 @@ namespace Singularity.Screen.ScreenClasses
             var road = toRemove as Road;
             var platform = toRemove as PlatformBlank;
             var settler = toRemove as Settler;
-            var controllableUnit = toRemove as ControllableUnit;
+            var freeMovingUnit = toRemove as FreeMovingUnit;
 
             if (!typeof(IDraw).IsAssignableFrom(typeof(T)) && !typeof(IUpdate).IsAssignableFrom(typeof(T)) && road == null && platform == null)
             {
@@ -455,9 +453,9 @@ namespace Singularity.Screen.ScreenClasses
             }
 
             // unsubscribe from this military unit when deleted
-            if (controllableUnit != null)
+            if (freeMovingUnit != null && freeMovingUnit.Friendly)
             {
-                mSelBox.SelectingBox -= controllableUnit.BoxSelected;
+                mSelBox.SelectingBox -= freeMovingUnit.BoxSelected;
             }
 
             if (typeof(IRevealing).IsAssignableFrom(typeof(T)))
