+++ Singularity/Singularity/Units/Target.cs
@@ -12,7 +12,6 @@ namespace Singularity.Units
         public Target(Vector2 position, Camera camera, ref Director director, ref Map.Map map) : base(position, camera, ref director, ref map)
         {
             mSpeed = 0;
-            mTarget = true;
         }
     }
 }
